Refactored custom-design to reduce the amount of code
